ai-facedetector-agent/requirements.txt
google-adk
google-generativeai
mediapipe
opencv-python
numpy
pillow
python-dotenv
typing-extensions